improved the nave bar on the pose and transition pages

+++ app/repo/Pose.scala
@@ -1,3 +1,3 @@
 package repo
 
-case class Pose(pose_id:Option[Int],name:String,created_by:String,image_url:Option[String],description_md:Option[String])
+case class Pose(pose_id:Option[Long],name:String,created_by:String,image_url:Option[String],description_md:Option[String])
